+++ CHANGELOG.md
@@ -4,12 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Added
+- Support for [referral event](https://developers.facebook.com/docs/messenger-platform/referral-params).
+
 ## [1.0.10] - 2017-03-27
-## Added
+### Added
 - Added `sendListTemplate` method.
 - Added `imageAspectRatio` option to `sendGenericTemplate` method.
 
-## Changed
+### Changed
 - Small change to how we call the `listeningAnswer` in Conversations to fix nexted `convo.ask()` not firing. (See #11).
 
 ## [1.0.9] - 2017-03-16
@@ -23,6 +27,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Added generic error message for when Messenger's API responds with an 'error' node.
 
+[Unreleased]: https://github.com/Charca/bootbot/compare/v1.0.10...HEAD
 [1.0.10]: https://github.com/Charca/bootbot/compare/v1.0.9...v1.0.10
 [1.0.9]: https://github.com/Charca/bootbot/compare/v1.0.8...v1.0.9
 [1.0.8]: https://github.com/Charca/bootbot/compare/v1.0.7...v1.0.8